split AppText and TitleText
