+++ src/components/AppText/index.js
@@ -1,10 +1,6 @@
 import React from 'react';
-import {Text, AppTitleText} from './style';
+import {Text} from './style';
 
 const AppText = props => <Text {...props}>{props.children}</Text>;
-
-export const TitleText = props => (
-  <AppTitleText {...props}>{props.children}</AppTitleText>
-);
 
 export default AppText;
