+++ src/components/TitleText/index.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import {Text} from './style';
+
+const TitleText = props => <Text {...props}>{props.children}</Text>;
+
+export default TitleText;
