+++ src/components/TitleText/style.js
@@ -1,5 +1,5 @@
 import styled from '@emotion/native';
 
 export const Text = styled.Text`
-  font-family: 'Inter';
+  font-family: 'Neo Sans Pro';
 `;
